Major Update!!! Fixing a lot of QoL stuff.
Updated the script to reflect some quality-of-life features.
First, I rearranged some of the line items to make them more readable. I also resized several of the boxes to make them better fit their contents. I also updated the naming scheme, allowing the use of a "quick rename" feature which appends the 7-digit serial number or service number to the end of the name. This is useful if you would like to keep the service number as part of the PC name. 
On request, I also updated the powershell window to open to a normal size during install, allowing updates to be clearly seen.
I also removed the "|out-host" piped command from the chocolatey installs, which should make them more easily readable.