Major Update to version 7.0!!!
This is the first major push in a few months, and with it come some graphical changes, a lot of cleaning up the interface order, a new renaming system, and a new "OK" function to close the initial form. 

First and foremost, my goal with this update was to simplify the "Quick Rename" function, making it easier to use, and more intuitive for the end user. The quick rename checkbox has now been replaced with a "Don't rename" checkbox, and two textboxes, one for regular renaming and one for quick renaming. This was done to simplify the instructions. I then use a series of If statements in their own function to make sure that the information in the boxes will work, including making sure only one is filled, and that it doesn't exceed the 15 character limit for a hostname. I also included a line to capitalize the quick-rename prefix to make it more readable.

After redoing the renaming system, I was left with lots of empty space, so I was able to reorganize the GUI just a bit, using less space, while still getting all the required information.

I also removed the option to install Teams with choco, as that has been causing issues as of late.

Going forward, I do plan on a graphical update to for aesthetics, but I haven't yet set aside the time for it.

I will also be updating the README to reflect some of the recent changes.