Added minizing/maximizing!
This is one of the things I have wanted since the beginning, and thanks to Nora Ballard's wonderful work at https://gist.github.com/Nora-Ballard/11240204 I was able to implement easy minimizing and maximizing of the shell window while the script is running. This can easily be adapted to hide instead of minimize, but I like the option of being able to open the window manually while the script is running.

Hopefully this is pretty simple to implement, but I'm going to create a new branch for now, just in case it breaks anything else in the script.